main: delay first packet to avoid resolve error
When first target is a hostname resolving will not complete before first
packet is to be sent, resulting in @ (resolving in progress). Delaying
will allow quick DNS servers or locally cached names to be looked up
before first packet is to be sent.

The 100ms delay should be fast enough to seemingly instant, but is it
still an abitrarily hardcoded value.