main: open raw socket only for xping
xping-unpriv and xping-http doesn't need raw sockets and should not
have a lingering socket that is privileged. This happens if root calls
xping-unpriv or xping-http directly.

Introduce a compile time option DO_SOCK_RAW enabling the privileged
socket call. Otherwise initialize the fd and errno indicating failure
to access a the SOCK_RAW.

Makefile is adjusted to compile with DO_SOCK_RAW only for xping.