main: makefile rules with all source variables
Most of the makefile rules had two seperate lists, which were almost
the same: the sources depended on and the stuff that should be linked
together.

Make the check-* dependencies empty c files and include them in the
linking process. Then switch to all source variables ($^ and $> for
gmake and pmake respectively) for specifying objects to be linked.
On gmake $> is empty and on pmake $^ is empty, including both ensures
building works with both.