main: silence unused result warnings for write()
Attempt to silence "warn_unused_result" from write(2) call for bell
characters.

If for some reason the bell character can't be printed, we can't do much
about it. Stopping the program isn't desired, neither is it likely we
can display an error message instead (if we can't write a single
character).