http: fix segfault when unable to connect
When the probe session is unable to connect, the event handler is
called from bufferevent_socket_connect(). The event handler in turn
calls session_free() which removes the session from the session list.

However LL_DELETE is not safe to call on an empty list (NULL pointer
dereference) and if this is the first session of the probe the list is
empty as connect happens before LL_APPEND.

This commit appends the session as soon as the session is associated
with a probe and the uses session_free in error handlers.